move Loading component to pages folder -> update the imported path for Loading component in components that used Loading
