+++ api/client/src/components/rightbar/RightbarInHomePage.js
@@ -1,5 +1,5 @@
 import { useGetFriends } from "../customHooks/useGetFriends";
-import Loading from "../Loading";
+import Loading from "../../pages/Loading";
 import OnlineFriends from "../online/OnlineFriends";
 import IconImage from "../profilePicture/IconImage";
